+++ NEWS.md
@@ -4,6 +4,7 @@
 
 # Version 0.2.0
 - Revision of trait parsing
+- New function sf2forest
 
 # Version 0.1.7
 - Compliance with medfate ver. 2.9.0